[Relocate] Base component for UI into UI package
