+++ Raemin/CloneCoding/app/src/main/java/com/example/deliveryclonecoding/ui/MainFragment.kt
@@ -1,16 +1,10 @@
 package com.example.deliveryclonecoding.ui
 
 import android.os.Bundle
-import android.util.Log
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import com.example.deliveryclonecoding.R
-import com.example.deliveryclonecoding.base.BaseFragment
+import com.example.deliveryclonecoding.ui.base.BaseFragment
 import com.example.deliveryclonecoding.databinding.FragmentMainBinding
-import com.example.deliveryclonecoding.databinding.FragmentSplashBinding
 
 
 class MainFragment : BaseFragment() {
